Moved virial radius code into separate file and added virial_test.
