+++ nbody/CMakeLists.txt
@@ -130,6 +130,7 @@ set(nbody_lib_src ${NBODY_SRC_DIR}/nbody_chisq.c
                   ${NBODY_SRC_DIR}/nbody_orbit_integrator.c
                   ${NBODY_SRC_DIR}/nbody_potential.c
                   ${NBODY_SRC_DIR}/nbody_bessel.c
+                  ${NBODY_SRC_DIR}/nbody_virial.c
                   ${NBODY_SRC_DIR}/nbody_density.c
                   ${NBODY_SRC_DIR}/nbody_friction.c
                   ${NBODY_SRC_DIR}/nbody.c
@@ -165,6 +166,7 @@ set(nbody_lib_headers ${NBODY_INCLUDE_DIR}/nbody_chisq.h
                       ${NBODY_INCLUDE_DIR}/nbody_orbit_integrator.h
                       ${NBODY_INCLUDE_DIR}/nbody_potential.h
                       ${NBODY_INCLUDE_DIR}/nbody_bessel.h
+                      ${NBODY_INCLUDE_DIR}/nbody_virial.h
                       ${NBODY_INCLUDE_DIR}/nbody_density.h
                       ${NBODY_INCLUDE_DIR}/nbody_friction.h
                       ${NBODY_INCLUDE_DIR}/nbody_check_params.h
