+++ nbody/src/nbody_lua_models.c
@@ -28,6 +28,7 @@
 #include "nbody_lua.h"
 #include "nbody_lua_types.h"
 #include "milkyway_lua.h"
+#include "nbody_virial.h"
 #include "nbody_plummer.h"
 #include "nbody_nfw.h"
 #include "nbody_hernq.h"
@@ -154,88 +155,6 @@ static int luaCalculateTimestep(lua_State* luaSt)
     return 1;
 }
 
-static real nbDwarfDensity(real r, real a_b, real a_d, real M_b, real M_d)
-{
-    /** FIXME: More dwarf profiles should be added here. For now, I just have Plummer spheres here. **/
-    real b_comp = M_b/mw_pow(a_b,3.0) * mw_pow(1.0 + r*r/a_b/a_b,-2.5);
-    real d_comp = M_d/mw_pow(a_d,3.0) * mw_pow(1.0 + r*r/a_d/a_d,-2.5);
-    return 3.0/4.0/M_PI*(b_comp + d_comp);
-}
-
-static real nbDwarfIntegral(real r, real a_b, real a_d, real M_b, real M_d)
-{
-    /** FIXME: More dwarf profiles should be added here. For now, I just have Plummer spheres here. **/
-    real b_comp = M_b/a_b * r * r * mw_pow(1.0 + r*r/a_b/a_b,-0.5);
-    real d_comp = M_d/a_d * r * r * mw_pow(1.0 + r*r/a_d/a_d,-0.5);
-    return 1.0/4.0/M_PI*(b_comp + d_comp);
-}
-
-static real nbCalculateVirial(real a_b, real a_d, real M_b, real M_d) /** General double integral used to calculate Henon length units (SPHERICAL PROFILES ONLY!)**/
-{
-    real M_tot = M_b + M_d;
-    real weight[5];
-    weight[0] = 0.2369268850561891;
-    weight[1] = 0.4786286704993665;
-    weight[2] = 0.5688888888888889;
-    weight[3] = 0.4786286704993665;
-    weight[4] = 0.2369268850561891;
-
-    real point[5];
-    point[0] = -0.9061798459386640;
-    point[1] = -0.5384693101056831;
-    point[2] = 0.0;
-    point[3] = 0.5384693101056831;
-    point[4] = 0.9061798459386640;
-    const int n = 10000;
-    int i1, j1, i2, j2;
-    const real lowerlimit = 0.0;
-    real upperlimit, a1, b1, a2, b2, r1, r2, r_lower, integral1, integral2;
-
-    //Integral goes to from zero to infinity, so we set upper limit to 100 times larger scale radius
-    if (a_b > a_d)
-    {
-        upperlimit = 1000.0*a_b;
-    }
-    else
-    {
-        upperlimit = 1000.0*a_d;
-    }
-
-    const real width = (upperlimit-lowerlimit)/(n*1.0);
-
-    if ((a_b <= 0.0)||(a_d <= 0.0))
-    {
-        mw_fail("Non-positive scale radius detected!");
-    }
-    else if ((M_b < 0.0)||(M_d < 0.0))
-    {
-        mw_fail("Negative mass detected!");
-    }
-    else if ((M_b == 0.0)&&(M_d == 0.0))
-    {
-        mw_fail("No dwarf masses detected!");
-    }
-    else
-    {
-        integral1 = 0.0;
-        for (i1 = 0; i1 < n; i1++)
-        {
-            a1 = lowerlimit + i1*width;
-            b1 = lowerlimit + (i1+1)*width;
-            for (j1 = 0; j1 < 5; j1++)
-            {
-                r1 = width*point[j1]/2.0 + (a1+b1)/2.0;
-                integral1 += weight[j1]*nbDwarfDensity(r1, a_b, a_d, M_b, M_d)*nbDwarfIntegral(r1, a_b, a_d, M_b, M_d)*width/2.0;
-            }
-        }
-    }
-
-    real R_v = M_tot*M_tot/16.0/M_PI/M_PI/integral1;
-    mw_printf("[a_b, a_d, M_b, M_d] = [%.15f, %.15f, %.15f, %.15f]\n", a_b, a_d, M_b, M_d);
-    mw_printf("Dwarf Virial Radius = %.15f\n", R_v);
-    return R_v;
-}
-
 static real nbCalculateEps2(real nbody, real a_b, real a_d, real M_b, real M_d)
 {
     real beta = 1.0;                                  /** Tunable parameter for softening length **/
@@ -243,7 +162,7 @@ static real nbCalculateEps2(real nbody, real a_b, real a_d, real M_b, real M_d)
     real eps = r_v * 0.98 * mw_pow(nbody, -0.26);     /** Optimal softening length pulled from Athanassoula et al. 1998 **/
     real eps2 = sqr(eps)/beta;
     if (eps2 <= REAL_EPSILON) {
-        eps2 = 2.0*REAL_EPSILON;
+        eps2 = REAL_EPSILON;
     }
     return eps2;
 }
